+++ pkg/diagnostics/client.go
@@ -40,7 +40,7 @@ func (c *Client) Units(node string) (*UnitsHealthResponseJSONStruct, error) {
 		}
 		return &units, nil
 	default:
-		return nil, fmt.Errorf("HTTP %d error", resp.StatusCode)
+		return nil, httpResponseToError(resp)
 	}
 }
 
@@ -65,7 +65,7 @@ func (c *Client) Cancel() (*BundleGenericResponseJSONStruct, error) {
 		}
 		return nil, errors.New(cancelBundle.Status)
 	default:
-		return nil, fmt.Errorf("HTTP %d error", resp.StatusCode)
+		return nil, httpResponseToError(resp)
 	}
 }
 
@@ -90,10 +90,13 @@ func (c *Client) Create(nodes []string) (*BundleCreateResponseJSONStruct, error)
 		return &createdBundle, err
 	case 409:
 		return nil, fmt.Errorf("another bundle already in progress")
-	case 503:
-		return nil, fmt.Errorf("Requested nodes %v not found", nodes)
 	default:
-		return nil, fmt.Errorf("HTTP %d error", resp.StatusCode)
+		var createdBundle BundleCreateResponseJSONStruct
+		err = json.NewDecoder(resp.Body).Decode(&createdBundle)
+		if err != nil {
+			return nil, httpResponseToError(resp)
+		}
+		return nil, fmt.Errorf("unexpected status code %s: %s", resp.Status, createdBundle.Status)
 	}
 }
 
@@ -126,7 +129,7 @@ func (c *Client) Delete(bundle string) (*BundleGenericResponseJSONStruct, error)
 	case 404:
 		return nil, fmt.Errorf("Bundle '%s' not found, unable to delete it", bundle)
 	default:
-		return nil, fmt.Errorf("HTTP %d error", resp.StatusCode)
+		return nil, httpResponseToError(resp)
 	}
 }
 
@@ -144,7 +147,7 @@ func (c *Client) List() (map[string][]Bundle, error) {
 		err = json.NewDecoder(resp.Body).Decode(&list)
 		return list, err
 	default:
-		return nil, fmt.Errorf("HTTP %d error", resp.StatusCode)
+		return nil, httpResponseToError(resp)
 	}
 }
 
@@ -165,6 +168,10 @@ func (c *Client) Status() (map[string]BundleReportStatus, error) {
 		}
 		return status, nil
 	default:
-		return nil, fmt.Errorf("HTTP %d error", resp.StatusCode)
+		return nil, httpResponseToError(resp)
 	}
+}
+
+func httpResponseToError(resp *http.Response) error {
+	return fmt.Errorf("unexpected status code %d", resp.StatusCode)
 }